demos/requirements.txt
ipympl==0.5.8
ipywidgets==7.5.1
jupyterlab==2.2.8
matplotlib==3.3.2
numpy==1.18.5
opencv_python==4.4.0.44
Pillow==6.2.2
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.5.2
tensorflow-gpu==2.3.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.32.2
transformers==2.9.0